pipeline_classes/requirements.txt
absl-py==1.1.0
aiofiles==22.1.0
aiohttp==3.8.1
aiosignal==1.2.0
aiostream==0.4.5
antlr4-python3-runtime==4.9.3
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==22.1.0
autocorrect==2.6.1
backcall==0.2.0
base58==2.1.1
bayesian-optimization==1.3.1
beautifulsoup4==4.11.1
bert-serving==0.0.1
bert-serving-client==1.10.0
bert-serving-server==1.10.0
bitarray==2.6.0
black==22.3.0
bleach==5.0.0
blis==0.9.1
cachetools==5.2.0
camelot-py==0.9.0
catalogue==2.0.8
certifi==2022.6.15
cffi==1.15.0
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
clip-client==0.7.0
clip-server==0.7.0
cloudpickle==2.2.0
colorama==0.4.4
combo==0.1.3
commonmark==0.9.1
confection==0.0.1
cPython==0.0.6
cryptography==38.0.1
cycler==0.11.0
cymem==2.0.6
Cython==0.29.28
cytoolz==0.12.0
dataclassy==0.11.1
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
-e git+https://github.com/facebookresearch/detectron2.git@0703e08a5f589f7503a3fbfce41309c80204eec8#egg=detectron2
distro==1.7.0
docarray==0.16.5
docker==6.0.0
eip712==0.1.0
Elixir==0.7.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
entrypoints==0.4
et-xmlfile==1.1.0
eth-abi==2.2.0
eth-account==0.5.9
eth-brownie==1.19.1
eth-event==1.2.3
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
execnet==1.9.0
executing==0.8.3
fairscale==0.4.10
fastapi==0.85.0
fastjsonschema==2.15.3
filelock==3.8.0
Flask==2.2.2
flatbuffers==2.0.7
fonttools==4.33.3
frozenlist==1.3.1
ftfy==6.1.1
funcy==1.17
future==0.18.2
fvcore==0.1.5.post20220512
gast==0.4.0
gensim==4.2.0
ghostscript==0.7
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.47.0
grpcio-health-checking==1.47.0
grpcio-reflection==1.47.0
h11==0.13.0
h5py==3.7.0
hexbytes==0.2.2
httptools==0.5.0
huggingface-hub==0.9.1
hydra-core==1.2.0
hypothesis==6.27.3
idna==3.3
imageio==2.22.3
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata==4.12.0
inflection==0.5.0
iniconfig==1.1.1
iopath==0.1.9
ipfshttpclient==0.8.0a2
ipykernel==6.13.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jcloud==0.0.35
jedi==0.18.1
jina==3.9.2
jina-hubble-sdk==0.17.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keras==2.10.0
Keras-Preprocessing==1.1.2
keybert==0.6.0
kiwisolver==1.4.2
langcodes==3.3.0
lazy-object-proxy==1.7.1
libclang==14.0.1
lime==0.2.0.1
llvmlite==0.39.1
lru-dict==1.1.8
lz4==4.0.2
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
missingno==0.5.1
mistune==0.8.4
multiaddr==0.0.9
multidict==6.0.2
murmurhash==1.0.8
mypy-extensions==0.4.3
mythx-models==1.9.1
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
netaddr==0.8.0
networkx==2.8.8
ninja==1.10.2.3
nltk==3.7
notebook==6.4.12
numba==0.56.2
numexpr==2.8.3
numpy==1.22.4
oauthlib==3.2.0
omegaconf==2.2.3
open-clip-torch==2.0.2
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parsimonious==0.8.1
parso==0.8.3
pathspec==0.9.0
pathy==0.6.2
patsy==0.5.2
pbr==5.10.0
pdf2image==1.16.0
pdfminer.six==20220524
pdfminer3==2018.12.3.0
pdfplumber==0.7.4
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.2
pluggy==1.0.0
portalocker==2.5.1
preshed==3.0.7
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.5
psaw==0.1.0
psutil==5.9.1
pure-eval==0.2.2
py==1.11.0
py-solc-ast==1.2.9
py-solc-x==1.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.5
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.9.2
pydot==1.4.2
Pygments==2.12.0
pygments-lexer-solidity==0.7.0
pyitlib==0.2.2
PyJWT==1.7.1
pyLDAvis==3.3.1
pymongo==4.1.1
PyMuPDF==1.20.2
pyod==1.0.5
pyparsing==3.0.9
PyPDF2==2.10.9
pyrsistent==0.18.1
pytesseract==0.3.10
pytest==6.2.5
pytest-forked==1.4.0
pytest-xdist==1.34.0
python-dateutil==2.8.1
python-dotenv==0.16.0
python-multipart==0.0.5
pythx==1.6.1
pytz==2022.1
PyWavelets==1.4.1
pywin32==304
pywinpty==2.0.5
PyYAML==5.4.1
pyzmq==23.1.0
qtconsole==5.3.1
QtPy==2.1.0
regex==2022.7.25
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
rlp==2.0.1
rsa==4.8
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
semantic-version==2.8.5
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
shap==0.41.0
six==1.16.0
sklearn==0.0
slicer==0.0.7
smart-open==5.2.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.4.2
srsly==2.4.4
stack-data==0.2.0
starlette==0.20.4
statsmodels==0.13.2
suod==0.0.8
tabula-py==2.5.1
tabulate==0.8.10
Tempita==0.5.2
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
textblob==0.17.1
thinc==8.1.1
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.6.7
tinycss2==1.1.1
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.1
torchaudio==0.12.1+cu113
torchvision==0.13.1
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
transformers==4.21.3
typer==0.4.2
typing_extensions==4.2.0
urllib3==1.26.11
uvicorn==0.18.3
varint==1.0.2
vvm==0.1.0
vyper==0.3.6
Wand==0.6.10
wasabi==0.10.1
watchfiles==0.17.0
wcwidth==0.2.5
web3==5.30.0
webencodings==0.5.1
websocket-client==1.4.1
websockets==9.1
Werkzeug==2.2.2
widgetsnbextension==3.6.0
wordcloud==1.8.2.2
wrapt==1.14.1
xgboost==1.6.2
xlrd==0.7.1
xlwt==0.7.2
yacs==0.1.8
yarl==1.8.1
zipp==3.8.1
